Change the definition of the ServerDetails struct key
Use only the serf Name for now.  Leaving the plumbing for now.
